Major refactor

+++ .zlogin
@@ -39,7 +39,7 @@
 
   # cleanup
   find $HOME -type f -name "*.zwc.old" -maxdepth 1 -delete
-  if [[ "$HOME" != "$ZDOTDIR" ]]; then
+  if [[ "$HOME" != "$ZDOTDIR" ]] && [[ -d "$ZDOTDIR" ]]; then
     find $ZDOTDIR -type f -name "*.zwc.old" -delete
   fi
 ) &!
